building: isolate the search for dynamic libraries
Isolate the search for dynamic libraries in a subprocess, and have
it import all collected packages in order to properly set up
library search paths. This functionality is implemented as a
module-level helper function that is executed via the new isolated
framework.

We need a stand-alone helper function because one of the tests needs
to monkey-patch it in order to inject its (subproess-)local override
of the getImports() function.