tests: fix test_regression::test_issue_5131
This test monkey-patches bindepend.getImports() function to ignore
errors on invalid binary file that the test uses. However, as we
now process binaries in an isolated subprocess, this change is not
progpagated there. So we need to provide getImports() override both
for the test's main process, as well as override the implementation
of build_main.find_binary_dependencies() to inject getImports()
override in the isolated subprocess.