hookutils: qt: fix DLL search path for PySide2 and PySide6
The DLL search path used to resolve import names to fullpaths on
Windows differs between PyQt5/6 and PySide2/6. For the former,
the existing BinariesPath is correct, but for the latter, we need
to use PrefixPath.

Alternatively, we could do away with giving extra search path to
bindepend.getfullnameof(), and instead perform the analysis of
Qt modules' DLL dependencies in an isolated subprocess and with
the main Qt package imported.

But at least for PyPI wheels, giving the search path seems to be
sufficient...