bootloader: apple events: improve forwarding of initial Apple event
In 2bf2d1d8ccfeee966ff22ae9432acd656f295a59, we introduced repeated
attempts at forwarding the initial Apple events in onefile .app
bundles, where the parent process might attempt to fowards received
events to the child process that is not ready to receive the events
yet.

The chosen time window of 5 seconds appear to be insufficient, leading
to sporadic event-forwading test failures on our CI.

Therefore, instead of attempting to forward the first event for
a limited time period, we now attempt to forward it forever, as
long as the child process is alive. This effectively suspends
subsequent events processing/forwarding, with the rationale that
until we manage to forward the very first event (i.e., the child
process becomes able to receive them), it is unlikely that we
would be able to successfully forward subsequent events, either.