macOS: robustify macOS assembly pipeline (#6188)
* osxutils: robustify the fix_exe_for_code_signing

When adjusting the size of the SYMTAB section and its parent
__LINKEDIT segment, directly compute the new size as the difference
between end-of-file and the corresponding offset. Previously, we
reconstructed the old file size from the __LINKEDIT segment's start
offset and size, and then computed the size delta based on the
new file size.

The new method is more direct and follows the basic assumption
that in the final file with appended CArchive, the SYMTAB section
and the __LINKEDIT segment must cover everything up to the end of
the file.

It also gives us some resilience against the codesign utility on
Mac OS 10.13 producing bogus __LINKEDIT sizes when removing  the
signature from the executable.

* building: macOS: disable signature removal when modifying binaries

Disable signature removal before modifying headers of the collected
binaries. Doing so makes the process faster, and avoids potential
bugs in the codesign utility (e.g., on Mac OS 10.13).