Implement support for importlib.resources (#6195)
Implement basic resource reader that allows access to on-filesystem resources 
(data files) via importlib.resources or its importlib_resources back-port for 
python 3.8 and earlier.

* tests: add a test for importlib.resources (or importlib_resources)

The test is modeled after the existing test for pkg_resources, and
uses the same test package, pyi_pkg_resources_provider. It attempts
to test the following functions from importlib.resources:
* contents()
* is_resource()
* path()
* read_binary()
* read_text()
* files() and as_file()

In contrast to pkg_resources test, we do not test zipped egg variant
here. Even in the unfrozen script, there is inconsistent behavior;
for example, under python 3.9, importlib.resources.contents()
recursively lists all sub-directories of a package (but not the
files within them).

The test should pass in the frozen version for importlib.resources
(python 3.9 and later) but not for importlib_resources (python 3.8
and earlier). The former appears to properly detect the on-disk and
provides access to them, whereas the latter does not, resulting in
various errors.

In order to support both variants, we will need to implement a
custom resource reader for our FrozenImporter.

* tests: install importlib_resources on python < 3.9

* loader: implement basic resourec reader for importlib.resource

Enables access to on-filesystem resources via importlib.resources
(python 3.9 and later) or its importlib_resources back-port
(python 3.8 and earlier).

Seems to be required for some packages (certifi, sphinx) under
python 3.10.

The current implementation lacks support for embedded resources,
but as those are mainly just collected .pyc files, the implemented
on-fileystem support should cover most use cases (i.e., access to
data files).