bootloader: fix compilation under winlibs mingw toolchain
Suppress two format-truncation warnings in the windows-only
codepath, which end up treated as errors by the winlibs mingw
toolchain.

The warnings arise because we do not check the return value of
the snprintf() calls (which might truncate the message), but in
those particular instances, we do not care about truncation.
Therefore, explicitly suppress the warnings via pragmas instead of
trying to work around them.