building: Windows: ensure onedir executable has no embedded manifest
When assembling Windows executable, first remove all resources from
the copy of bootloader file. This ensures that executable contains
no embedded manifest, even if the bootloader was compiled using a
toolchain that forcibly embeds a default manifest (e.g., mingw gcc
toolchain from msys2).

In onefile builds, the manifest (if already present) is rewritten
later on, so it poses no problem. In onedir builds, however, an
already present manifest makes Windows ignore the generated
external manifest. This means that options such as uac-admin are
effectively ignored, and that Common Controls v6 are not loaded
for windowed bootloaders, resulting in an entry point for
LoadIconMetric not being found.