bootloader: exception dialog: use %ls instead of %s with swprintf
It seems that with MinGW-w64 gcc, the %s format argument in a
wide-char swprintf() function call does not correspond to wide-char
(as it does under MSVC), but rather to single-char. Therefore, with
mingw-compiled bootloader, the exception dialog displays only initial
letters of script name and exception message.

Use %ls to remove all ambiguity.