tests: add a test for checking sys.executable w.r.t. append_pkg
Add a test that verifies the (basename of the) sys.executable for
all combinations of onefile/onedir and append_pkg EXE() parameter
(True = embedded CArchive, the default; False = sideloaded
CArchive).

The sideload test variant currently fails in both onedir and
onefile mode. In onedir mode, the sys.executable ends up pointing
to the pkg file instead of the executable. When building in onefile
mode, the pkg file is created next to the .spec file instead of next
to the executable.