building: fix location of the generated stand-alone PKG file
Fix the location of the generated stand-alone PKG file to be used
in the sideload mode (enabled via append_pkg=False argument to the
EXE().

The existing implementation creates the PKG file in the workdir, as
build/program/PKG-01.pkg. If append_pkg=False, this file is then
copied to program.pkg (in current working directory / next to .spec).
Then, in onedir mode, the program.pkg is copied (by COLLECT() target)
to dist/program/program.pkg. In onefile mode, it is left in the
current working directory, whereas it should be placed next to the
executable, in dist/program.pkg. Furthermore, the EXE() guts check
tests for existence of the program.pkg in current working directory.

This commit changes the name of initial .pkg file; it is now
generated as build/program/program.pkg, and EXE() guts check tests
for existence of that file. If append_pkg=False, this file is then
copied next to the executable - in onefile mode, we do that
explicitly (in place of the code that used to copy it to cwd),
whereas in onedir mode, we let COLLECT() take care of that (same
as before).

In onefile mode, the stand-alone PKG is now correctly placed next
to the executable by default. In onedir mode, an extra copy step
has been eliminated.