building: Windows: do not collect copy of a manifest in onefile bundle
Stop collecting a copy of the manifest in the onefile bundle, and
stop passing it to bootloader via the pyi-windows-manifest-filename
bootloader flag.

Ever since #3746 introduced manifest embedding for onefile builds,
the frozen application has the default activation context available
(created and activated by the presence of the embedded manifest).
As part of an earlier manifest-handling codepath, the bootloader
attempts to create its own activation context with the manifest
given via the pyi-windows-manifest-filename bootloader flag.
But as the default activation context is already present, this
attempt fails, resulting in error: "An attempt to set the process
default activation context failed because the process default
activation context was already set."

While the error is non-fatal (and visible only in debug mode), it
can be very confusing for users who are trying to debug their frozen
application.

So as the activation-context codepath is basically a no-op, we
can stop using the pyi-windows-manifest-filename bootloader flag,
and stop bundling the copy of the manifest in onefile builds. And
we can probably remove the activation-context codepath from the
bootloader as well.