bootloader: Fix searching for 'strip'
Previously the waf config would search for 'strip' only in the same
directory as the compiler. This breaks on systems which have custom
compilers, at the least.

Instead, search for 'strip' using the path to the compiler as a
suggestion, but still look elsewhere on the normal PATH as well.

This change also affects searching for 'windres' on Windows systems
using non-MSVC compilers. It is not clear whether this is used, but the
same logic should work.