tests: add a test for str representation of VSVersionInfo
A basic test that "serializes" the VSVersionInfo struct to string,
and then "deserializes" it back via an eval(). This simulates the
serialization done by the grab_version.py utility when writing
the version info to a text file, and deserialization done by
the versioninfo.SetVersion() function when reading the info
from the said file.