winutils: versioninfo: fix the declared string length in StringStruct
The second entry (wValueLength) in the String struct denotes
"the size, in words, of the Value member." This means that it
corresponds to the number of (wide-)characters in the string, and
not the number of bytes.

Fixes the declared string lengths that were twice too long, and
thus their retrieval resulted in trailing garbage characters
(usually not visible in C-based programs due to the terminating
NULL character, but visible when read in python via ctypes and
winver API).