building: reorganize the EXE assembly function
Reorganize the EXE assembly function so that its operations are
grouped into three categories/steps:
1. bootloader copy & pre-processing
2. appending the PKG data
3. post-processing

This ensures that the pre-processing and post-processing steps are
always performed, regardless of whether PKG is appended to the
executable or not.

Fixes #6235, which is caused by pre-processing and post-processing
not being peformed on macOS when using stand-alone PKG (i.e., the
append_pkg=False option to EXE()).