building: Windows: embed manifest into onedir executable by default (#6243)
Embed the manifest in onedir executable by default, in order to
avoid potential issues when user renames the executable. For example,
the manifest not being found anymore due to activation context caching
when user renames the executable and attempts to run it before also
renaming the manifest file.

If the old behavior of generating an external manifest file is
required for some reason (e.g., manual modification in a
post-processing step), it can be re-enabled using the new
--no-embed-manifest command-line switch or by setting the new
embed_manifest argument to EXE() to False.