Cleanup: apply missing reformatting and line rewrap
Couple of files on this branch seem to have been exempted from
reformatting and line rewrap - presumably due to a merge conflict
when the reformatting commit was back-ported from develop branch.

This commit synchronizes the formatting in these files so that
it is consistent with those in develop branch, while preserving
functional differences that have accumulated since the branching.