building: makespec: do not hard-code spec file's directory as pathex
Do not hardcode the spec file's parent directory as pathex in the
generated spec file.

Doing so makes the spec file appear non-portable, as the
hard-coded absolute path is both OS-specific and depends on
current location of the spec file.

In practice, the spec file's location should also have little
relevance to the module search path. The spec file is either
located in a different directory than the entry-point script(s),
which makes its location irrelevant. Or, it is located in the
same directory as the entry-point script(s), in which case the
location is dynamically added to pathex due to Analysis calling
its _extend_pathex() method.