Bindepend: Skip querying ldconfig's cache on musl.
Musl does not use ldconfig cache at all. The ldconfig executable, strictly
speaking, is not supposed to exist on a traditional musl system but a dummy
executable which complains if it receives any arguments may be put in its place
for compatibility. Running this dummy should be avoided because it writes a
visible error message to stderr.

Update the corresponding test to reflect that LDCONFIG_CACHE will always be
empty on musl.