Update towncrier to >= 21.3.0.
Said version made a breaking change - to move the version headings into the
jinja2 template so that anyone using a custom template (like us) is short of a
title. This commit adds the title to our own custom template and pins the
towncrier verion to >= 21.3.0 so that we don't accidently get two titles.

See https://github.com/twisted/towncrier/pull/147 and
https://github.com/twisted/towncrier/issues/340 for details.