compat: extend is_macos_11 flag to cover macOS 11 or later
On macOS 12 Monterey, platform.mac_ver()[0] returns '10.16' if in
compatibility mode (i.e., running python without Big Sur support)
or '12.0.1' if in native mode (python with Big Sur support).

Therefore, extend the is_macos_11 flag to signify that we are
running on macOS 11 or later (instead of just macOS 11).

This ensures that problematic tests are skipped on macOS 12 Monterey,
same as they are on macOS 11 Big Sur:
* test_ctypes_util_find_library if we are running in compatibility
  mode, in which ctypes.util.find_library() does not find system
  libraries
* test_apple_event_handling_carbon if pytest's base temporary
  directory is in the default location (/var), in which case the
  custom URL and file schema registration does not work for app
  bundles.