winmanifest: fix structure of generated dependentAssembly entry
If we do not obtain dependent assembly entry for MS Common Controls
v6 from dependency analysis on python executable, we manually
add an entry for it. However, this mechanism currently generates
invalid entry that looks like this:

```xml
  <dependency>
    <dependentAssembly>
      <assemblyIdentity type="win32" name="Microsoft.Windows.Common-Controls"
                        language="*" processorArchitecture="amd64" version="6.0.0.0"
                        publicKeyToken="6595b64144ccf1df">
      </assemblyIdentity>
      <compatibility xmlns="urn:schemas-microsoft-com:compatibility.v1">
      </compatibility>
      <application xmlns="urn:schemas-microsoft-com:asm.v3">
      </application>
    </dependentAssembly>
  </dependency>
```

Windows does not allow `application` tag within `dependentAssembly`
(anymore), leading to the following error when the manifest is
embedded in the executable:
```
The application has failed to start because its side-by-side
configuration is incorrect. Please see the application event log or
use the command-line sxstrace.exe tool for more detail.
```

Checking the output of `sxstrace` displays the following error:
```
The element application appears as a child of element
urn:schemas-microsoft-com:asm.v1^dependentAssembly which is not
supported by this version of Windows.
```

So prevent both `application` and `compatibility` tags from leaking
into `dependentAssembly`.