bootloader: Windows: fix behavior when runtime tmpdir is a drive letter
Fix extraction behavior of Windows onefile builds when the run-time
temporary directory (--runtime-tmpdir) is set to just a drive letter
(e.g., "c:"), either directly or via an expanded environment variable,
such as %HOMEDRIVE%.

In such cases, we must avoid calling _wfullpath, because as per
its docs, "If the relPath argument specifies a disk drive, the
current directory of this drive is combined with the path". Instead,
we should treat such paths separately, so that the final temporary
directory is created in the drive's root (e.g., "c:\_MEIXXXXX").