hooks: add support for Anaconda-installed PyQtWebEngine on macOS
Add support for Anaconda-installed PyQtWebEngine (Qt5) on macOS. In
contrast to PyPI wheels that ship framework-based Qt5 build,
the Anaconda's Qt5 build is dylib-based and thus similar to Linux
and Windows builds. Therefore, the following changes are needed:

1. modify the get_qt_webengine_binaries_and_data_files helper to
identify non-framework-based Qt build on macOS, and divert the
execution to shared-lib-based (Windows, Linux) codepath.

2. modify PyQt5.QtWebEngine runtime hook to set the
QTWEBENGINEPROCESS_PATH environment variable only if the
QtWebEngineProcess executable was collected from a framework-based
build (this is determined by its location, or rather, its existence
in pre-determined location). In a dylib-based build, the executable
is discovered without having to set the environment variable, same
as on Windows and Linux. The PySide2 hook is also modified in the
same way to keep them in sync.

3. modfy the get_qt_webengine_binaries_and_data_files helper to
collect the QtWebEngineProcess executable as a binary file instead
of a data file. On Windows and Linux, this makes no difference, but
on macOS it ensures that the library paths are properly rewritten
so that the helper executable can find the collected Qt5 libs.