osxutils: fix codesign regression in construction of universal2 executables
131c4dc broke universal2 support, because the simplified
computation of section/segment length fails to account for the
fact that their starting offsets are not relative to start of the
file, but rather to the start of the arch slice. In thin binaries,
the start of slice corresponds to the start of file, but in fat
binaries, we need to account for the non-zero header.offset that
contains the start offset of the slice.

Due to incorrectly computed section and segment sizes, the resulting
universal2 executable failed to pass strict codesign validation.