building: fix error when a linux .so library is collected on macOS
The .so is a valid binary suffix on both linux and macOS, which
means that a linux .so library may end up collected as a binary
on macOS (e.g., if a wheel provides shared libraries for all OSes
and collect_dynamic_libs or collect_all is used to collect files
from the package).

Trying to run macOS-specific processing on linux shared libraries
results in errors, so we need to detect an incompatible binary
as soon as possible and avoid attempts at any further processing,
so that the file is collected as-is.