bootloader: normalize the caught exception before processing it
When the user script terminates with uncaught exception and we
retrieve its data via PyErr_Fetch(), we must also normalize the
retrieved data via PyErr_NormalizeException() before trying to
process it further.

Otherwise, the retrieved value instance may not be an instance of
the indicated class object, which results in an error when we try
to format traceback via traceback.format_exception(). Consequently,
we fail to retrieve traceback to display in the windowed
bootloader's error reporting facility (i.e., the uncaught exception
dialog on Windows; syslog on macOS).