hooks: qt: avoid collecting the QtQml module when not necessary
Avoid collecting the whole QtQml module when there is no indication
that the frozen application is actually using QtQml, such as the
import of the QtQuick* module(s) or the QtQml module itself.

On linux, the PySide2 and PySide6 extension modules seem to be
always linked against libQt5Core, libQt5Network, and libQt5Qml
shared library (or their Qt6 equivalent), which means that we end
up collecting QtQml even for the basic widget-based application.
And because we collect all QtQtml data and plugins, we end up
pulling in several additional Qt libraries and modules as well.

Similarly, the QtWebEngine* extension modules of all four Qt-based
bindings (PyQt5, PyQt6, PySide2, and PySide6) on all OSes seem to
be linked against QtQml shared library as well, even though the
module can be used in a pure widget-based application as well.

Therefore, prevent the linked libQt5Qml/libQt6Qml shared library
from triggering the collection of the QtQml module, except for
the modules whose names start with either QtQml* or QtQuick*.