bootloader: refactor the PKG cookie search code
Refactor the PKG MAGIC pattern search code into a generic helper
function, so that we can reuse it elsewhere.

Also rework the way in which we prevent the PKG MAGIC pattern to
be stored in the executable; instead of having it stored in two
parts with extra data in between, store a base pattern with fourth
byte zeroed out. At the start of search function, create a copy
and set the zeroed-out byte to correct value. The search function
itself can then directly use this buffer, with a single memcmp()
call.