bootloader: lock the PKG sideload mode
Refuse to side-load an external PKG unless the executable has a
special signature embedded.

The aim is to ensure that the executable used by sideload-enabled
applications in the wild is at least slightly different from the stock
pre-built bootloaders that we ship in our wheels. This in turn should
help against hash-based associations made by the AV software.