tests: skip tkinter-based tests if the module is not importable
Instead of having an xfail mark for tkinter-based tests on macos,
check if tkinter is importable, and if not, skip the tests.

On macos-10.15 runners these tests have been XPASSing, and due to
missing tcl/tk 8.6 shared libraries on macos-11 runners, they have
been XFAILing (whereas prior to introduction of python 3.10.1,
they have been XPASSing as well).

Extend this to idlelib test as well, because it depends on tkinter
(and is now failing under python >= 3.10.1 on macos-11).