Fixed pkgutil.iter_modules and improved the code a bit
The old implementation would only work if the path started with the _MEIPASS directory and would throw an assertion error if it did not.
This version only runs the pyinstaller code if the path starts with _MEIPASS and allows other paths to work.
It also supports multiple paths and correctly handles an empty list.
I have also cleaned up the code a little to make it more readable.