hooks: fix pkgutil.iter_modules when paths involve symbolic links
In our pkgutil.iter_module replacement, we need to fully resolve
both the sys._MEIPASS and the given search paths, in case either
contains a symbolic link. Failing to do so leads to path mis-match
when symbolic links are involved and the given search paths are
already fully resolved. This may happen on macOS with onefile
builds if the caller of pkgutil.iter_module fully resolves the
search path(s) before calling the function (issue #6537).

Add a test that reproduces scenario from #6537.