hooks: add support for setuptools 60.7.1, mark 60.7.0 as unsupported
As of 60.7.0, setuptools vendored jaraco.text package in its
pkg_resources, and their VendoredImporter does not seem to play
nicely with our FrozenImporter when we collect the
"Lorem Ipsum.txt" data file that is required by the vendored
jaraco.text package.

In 60.7.1, setuptools dropped the "Lorem Ipsum.txt" data file from
the vendored jaraco.text package, so we can accommodate it with
a couple of hidden imports.

Initially, we also had a work-around for 60.7.0, but that worked
only because it accidentally de-vendored the jaraco.text package,
and may not play nicely in situations when the application also
uses the actual jaraco.text package. So it is easier and safer to
exit with an error in case setuptools 60.7.0 is encountered, and
tell user to perform either downgrade or upgrade.