Building: Replace .exe suffix of temporary EXE files (#6467).
Whenever an intermediate/temporary executable file needs to be written, give it
the filename extension .notanexecutable instead of .exe. This fools Microsoft
Defender into thinking that the file is not an executable which requires
scanning. Even though the scan does not quarantine the EXE, its being opened and
read by Defender means that it can't be opened in write mode by a later step in
the executable assembly process.

I'm choosing to add this suffix on all platforms despite it only being needed on
Windows so as to minimise platform dependent logic.