makespec: fix regression w.r.t. default onefile and console value
If onefile- and console-related options are not specified on the
command line, the corresponding values are now None and we need
to explicitly convert them to their default values (onefile=False
and console=True). Otherwise, we end up with (onefile=False and
console=False), which is a regression in behavior.