splash: use bundled vcruntime140.dll in Windows onefile builds
On Windows, add vcruntime140.dll to the list of splash requirements
in order to have it extracted into the initial runtime directory
when running in onefile mode. This way, when we try to load Tcl/Tk
shared libraries, the extracted copy of vcruntime140.dll is used
instead of system-installed copy (which may be unavailable if
the system does not have VC runtime installed).