Issue error when `makespec` arguments are passed, but `makespec` will not be invoked (#6660)
* Issue error when makespec args are passed with a .spec file

* test utils: fix default arguments used by AppBuilder

Prevent the AppBuilder (used by pyi_builder test fixtures) from
setting up arguments from the makespec group when testing against
a spec file. Trying to do so now results in an error.

* makespec: fix regression w.r.t. default onefile and console value

If onefile- and console-related options are not specified on the
command line, the corresponding values are now None and we need
to explicitly convert them to their default values (onefile=False
and console=True). Otherwise, we end up with (onefile=False and
console=False), which is a regression in behavior.

Co-authored-by: Rok Mandeljc <rok.mandeljc@gmail.com>