tests: add subtraction without typecodes to TOC unit tests
The added tests show that we cannot consistently implement Windows
specific case-insensitive comparison behavior when performing
subtraction without typecode, due to ambiguity. Either we do not
perform case normalization for entries without typecode, and break
the case-insensitive behavior for DATA and BINARY entries; or we
perform case normalization, and break the case-sensitive behavior
for PYMODULE and other entry types.