hookutils: add support for resolving metadata .egg-info files
Debian and Ubuntu linux seem to often package python packages
with metadata stored in an .egg-info file as opposed to .egg-info
directory or the standard .dist-info directory. In such cases,
pkg_resources.get_distribution() returns an instance of
pkg_resources.EggInfoDistribution that has egg_info field set
to None, and we need to attempt to resolve the egg-info path
ourselves.

This commit effectively restores a fallback that was in place
before #5774, but adds an additional search pattern based on
project_name and version to supplement the original pattern
based on egg_name().