Fix when the entry-point script and a package have the same name. (#6711)
* Fix when the entry-point script and a package have the same name.

When the entry point script's bytecode is collected before archiving, the
bytecode of the __init__.py of a package with the same name would be collected
instead if it exists. This mix-up can be avoided simply by removing the complex
pkgutil shenanigans used to get the entry-point's bytecode in a way that makes
use of pycache. Given that there is rarely more than one entry-point script, this
loss of performance is insignificant.

Fixes #4771, fixes #6702.