Make --add-data/--add-binary path separator platform independent. [skip ci]
`--add-data=source:dest` is now allowed (and preferred) on Windows as
well as UNIX. On Windows, the prior behaviour of using `;` is still
allowed to avoid breaking backwards compatibility. Special handling is
given to ignore colons at the start of the string, preceded by a single
letter and followed by a slash so as to avoid splitting
`--add-data=C:\foo\bar:dest` in the wrong place. Because `:` is an
illegal character in Windows filenames, this avoids being a breaking
change since no Windows users could have been `:` literally.