hooks: move qt.conf collection into QtWebEngine hooks
The qt.conf file is used by the helper QtWebEngineProcess executable
to locate the Qt shared libraries. Therefore, it makes more sense
to collect this file as part of QtWebEngine hooks than the base
PySide2/PyQt5/PySide6/PyQt6 hook; more so because in case where
QtWebEngine is provided by separate wheel (PyQt5, PyQt6), the
qt.conf file is actually provided by that wheel and not the base
one.