hooks: pickle: exclude doctest
Exclude doctest in pickle hook. As already noted there, it
is needed only when running the module as __main__ (same as
already-excluded argparse).

The doctest module pulls in several other modules that are
hidden imports in PySide and PyQt packages. Specifically:
- PySide2 and PySide6 require inspect module
- PySide2 prior to 5.15 requires __future__ module
- PyQt5 and PyQt6 require pkgutil module

So add these as hidden imports to corresponding hooks.