bindepend: get_python_library_path: resolve symlinks during libdir search
When searching for python shared library via libdir search (i.e,
using _find_lib_in_libdirs() helper), resolve the potential
symbolic links. This ensures that unversioned .so names (e.g.,
libpython3.x.so) are resolved as versioned .so names (e.g.,
libpython3.x.so.1), which is the behavior of linker-based search
codepath.

Having the bootloader expect an unversioned .so name results in
a missing library error when the versioned .so library is already
collected for whatever reason (e.g., due to being link-time
dependency of a collected extension), because such cases preclude
collection of the unversioned .so name.

In practice, the above issue manifests only under Anaconda on
linux, because regular linux codepath uses linker-based search,
which already resolves unversioned .so name to the versioned one.