hookutils: collect_submodules: do not recurse into filtered subpackages
Reorganize the collect_submodules code so that each (sub)package
scan step is performed separately, and the filter function is
applied to the results before recursing into the child subpackages.

This ensures that we do not attempt to recurse into subpackages
that are filtered out (i.e., packages that are known to be
un-importable and generate a warning/error depending on the
on_error setting).