hooks: add support for setuptools-provided distutils
Add support for setuptools-provided distutils in order
to allow frozen application to import setuptools without
having to force the use of stdlib-provided distutils
(i.e., setting SETUPTOOLS_USE_DISTUTILS environment
variable to something else than 'local').

At the same time, we want to prevent duplication that arises
from setuptools._distutils being collected as distutils as
well as under original name (the latter is also needed for
setuptools to be importable in the frozen application).

Hence the following hooks are needed:

1. pre-safe-import-module hook

Create a new pre-safe-import-module hook for distutils. If using
setuptools-provided distutils, it marks distutils as runtime
module. This inhibits any further analysis and prevents automatic
collection (which we need to prevent duplication). On the downside,
we need to mark all submodules as runtime modules as well. And,
we will need the standard hook to actually collect everything from
setuptools._distutils.

2. pre-find-module-path hook

Update the pre-find-module-path distutils hooks. Since this hook
is not ran if using setuptools-based distutils (bacause analysis
is inhibited thanks to pre-safe-import-hook marking distutils
as runtime module), it does not require any additions.

However, we do need to avoidd importing distutils and opcode
directly; use the hookutil get_module_file_attribute() to obtain the
__file__ paths.

Also, the virtualenv-version retargetting logic had a bug in
path comparison that caused it to peform retargetting in any case,
even when it was not necessary (when stdlib version of distutils
was used anyway).

3. standard hook

Modify the standard distutils hook to detect the use of
setuptools-provided distutils and collect all submodules of
setuptools._distutils. Also include _distutils_hack among
hidden imports, just in case.

4. runtime hook

Add a runtime hook for distutils that performs same function
as distutils-precedence.pth from the setuptools package. Namely,
it installs a special meta finder that diverts the import of
distutils to setuptools._distutils, if available.