tests: test_import_pep302: do not use distutils as test package
Swap `distutils` and `distutils.config` with `json` and `json.encoder`
in the `test_pep302_loader_frozen_{package,submodule}` tests.

The `distutils` is the one package that is not guaranteed to have
a loader with optional PEP302 functions implemented anymore; the
setuptools-provided version has a custom loader that does not
implement the optional functionality.