hooks: add pre-safe-import-module hook for gi
Add a pre-safe-import-module hook for `gi` that on linux checks
the `gi.__path__` and extends the modulegraph's paths for the
package.

The RHEL/Fedora RPM package for GObject introspection is known
to split the `gi` package into two locations:
 - /usr/lib64/python3.x/site-packages/gi
 - /usr/lib/python3.x/site-packages/gi
The `__init__.py` is located in the first directory, while
`repository` and `overrides` are located in the second, and
`__init__.py` dynamically extends the `__path__` using
`__path__ = pkgutil.extend_path(__path__, __name__)`.

The modulegraph has no way of knowing this, so we need to extend
the package paths via the pre-safe-import-module hook. Otherwise,
only first location is scanned, and the `gi.repository` ends up
missing.