tests: modulegraph unit tests: do not use distutils as test package
Swap `distutils`, `distutils.ccompiler`, and `distutils.sysconfig`
with `json`, `json.encoder`, and `json.decoder`, respectively, in
the `test_imports::TestRegressions3::testRegr1`.

The test scenario is not tied to the `distutils` per se; it
involves a custom package (`mypkg`) creating a subpackage that
essentially forwards to the stdlib package (e.g., `mypkg.distutils`
in the original version).

The test fails with `setuptools`-provided `distutils`, because
modulegraph fails to account for the meta-loader used by `setuptools`
to override `distutils`. Therefore, the modulegraph picks up the
stdlib version, while `import` statement picks up the `setuptools`
version, leading to error when comparing the paths.

As this problem is not tied to the test scenario and as the test
scenario is not tied to `distutils`, the easiest fix is to switch
to another stdlib package, in this case `json`.