depend: prevent get_code_using() from returning None code objects
In addition to all other checks and restrictions imposed in the
`analysis.get_code_using`, also check that code object is not
None.

This seems to happen in #6873, with a directory named `packaging`
placed next to entry point script seems to be treated as a namespace
package, and mistaken for a referrer of `pkg_resources`, whereas
the original referrer is likely the vendored version of `packaging`
from the `pkg_resources` itself.