hookutils: fix regression in get_gi_typelibs behavior
Due to the way the `get_gi_typelibs` function is used in the `gi`
hooks, it should not raise an exception when the package cannot
be found, but rather display a warning and return empty lists
(which is the behavior in PyInstaller 4.x). Otherwise, a missing
optional dependency causes build-time error and aborts the build.