hooks: refactor GObject introspection (gi) hooks
The modules imported from gi.repository are marked as runtime
modules by their corresponding pre-safe-import-module hooks.
Therefore, their standard hooks are always loaded and executed,
regardless of whether the modue is actually importable or not.

In PyInstaller v5, this behavior triggers errors in hooks for
GI modules that are not importable, because the new `isolated`
framework propagates the errors instead of swallowing them.
While these errors could be caught and demoted to warnings
to match the old behavior, it would be better hooks checked
whether module is importable before doing any processing
at all.

To that end, we introduce new class, `GiModuleInfo` that,
as part of its initialization, allows us to:
 - perform availability check
 - obtain data previously returned by `get_gi_typelibs`
 - obtain data previously returned by `get_gi_libdir`
using a single isolated import attempt (instead of one
being performed in each of those steps).

In  addition, if passed `hook_api` as an optional argument,
the `GiModuleInfo` can use hook configuration API to override
the GI module version to be collected (which allows the
standard use pattern to be removed from the hook itself).

The old `get_gi_typelibs` and `get_gi_libdir` functions
now internally use `GiModuleInfo` to provide backward
compatible behavior to (potential) exetnal user.

All `gi` hooks are ported to the `GiModuleInfo` and now
become no-op if the module is not available.

In addition, hooks are cleaned up/refactored so that all
processing is performed either in the loading stage ("simple"
hooks that do not require access to hook configuration API)
or in the `hook()` function (hooks that require access to
hook configuration API), but not in the mixture of the two.